Refactor configuration loading to use `config_path` utility for consistent and dynamic path management.
