+++ Forceringen/PLC/Search_Access.py
@@ -158,7 +158,7 @@ class DatabaseSearcher:
 
 
 if __name__ == "__main__":
-    from Forceringen.config.yaml_path import config_path
+    from Forceringen.config.config_path import config_path
     """
         Information:
             Main execution block that demonstrates the usage of the DatabaseSearcher class.
